ai_service/requirements.txt
Flask==2.3.3
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
tensorflow==2.13.0
opencv-python==4.8.1.78
roboflow==1.1.9
inference-sdk==0.9.10
